Generalize ShardReindexerV3 beyond Map→Blockmax
Extract strategy-specific logic from the monolithic
ShardReindexTask_MapToBlockmax into a MigrationStrategy interface,
enabling future migration types (filterable, rangeable, etc.) to
reuse the same lifecycle state machine.

- New MigrationStrategy interface for strategy-specific behavior
- New MapToBlockmaxStrategy implementing blockmax-specific logic
- New ShardReindexTaskGeneric with all lifecycle logic parameterized
- Slim inverted_reindexer_map_to_blockmax.go from ~2000 to ~90 lines
- Renamed tracker types to be strategy-agnostic (reindexTracker, etc.)
- Full backward compatibility preserved (disk layout, overrides, API)
